refactor: start working with login page
